+++ test/localization_test.dart
@@ -1,14 +1,17 @@
+import 'package:anas_localization/localization.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:anas_localization/src/core/localization_service.dart';
-import 'package:anas_localization/src/generated/dictionary.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
+  setUp(() {
+    LocalizationService().clear();
+    LocalizationService.clearPreviewDictionaries();
+    LocalizationService.supportedLocales = ['en', 'tr', 'ar'];
+  });
+
   group('LocalizationService', () {
     test('loads English dictionary and returns correct values', () async {
-      // Load the English locale
       await LocalizationService().loadLocale('en');
       final Dictionary dict = LocalizationService().currentDictionary;
 
@@ -16,16 +19,13 @@ void main() {
     });
 
     test('throws error for unsupported locale', () async {
-      expect(
-            () => LocalizationService().loadLocale('fr'),
-        throwsException,
-      );
+      expect(() => LocalizationService().loadLocale('fr'), throwsException);
     });
-  });
 
-  test('loads Turkish dictionary and returns correct values', () async {
-    await LocalizationService().loadLocale('tr');
-    final Dictionary dict = LocalizationService().currentDictionary;
-    expect(dict.welcome, equals('Hoş geldiniz'));
+    test('loads Turkish dictionary and returns correct values', () async {
+      await LocalizationService().loadLocale('tr');
+      final Dictionary dict = LocalizationService().currentDictionary;
+      expect(dict.welcome, equals('Hoş geldiniz'));
+    });
   });
 }